Split up implementation definitions
